+++ backend/src/main/java/com/bruno/artistalbum/repository/RegionalRepository.java
@@ -3,7 +3,7 @@ package com.bruno.artistalbum.repository;
 import com.bruno.artistalbum.model.Regional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RegionalRepository extends JpaRepository<Regional, Long> {
+public interface RegionalRepository extends JpaRepository<Regional, Integer> {
     java.util.List<com.bruno.artistalbum.model.Regional> findByAtivoTrue();
 
     java.util.Optional<com.bruno.artistalbum.model.Regional> findByIdAndAtivoTrue(Integer id);
